chore: Update XGClassifier_monthly_ai.py for occupancy level prediction at classification level - sitting on 94% accuracy, refining it
